example/requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
msgpack-python==0.2.4
wsgiref==0.1.2
